pyproject.toml
[tool.poetry]
name = "bot-udacha"
version = "0.2.0"
description = ""
authors = ["PavelPanchenko <pavel_panchenko@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.4"
aiogram = "2.24"
aioredis = "2.0.1"
fastapi = "0.89.1"
ormar = "^0.12.0"
uvicorn = "^0.20.0"
colorlog = "^6.7.0"
redis = "^4.4.2"
requests = "^2.28.2"
pillow = "^9.4.0"
aiogram-calendar = "^0.1.1"
jmespath = "^1.0.1"
python-dotenv = "^0.21.0"
asyncpg = "^0.27.0"
psycopg2 = "^2.9.5"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
